Avance del proyecto

+++ app/public/routes.py
@@ -40,6 +40,7 @@ def post(id):
     post=Posts.query.get_or_404(id)
     return render_template('public/post.html', post=post)
 @public_bp.route("/editPost<int:id>", methods=['GET','POST'])
+@login_required
 def editPost(id):
     form=PostForm()
     post=Posts.query.get_or_404(id)
@@ -51,22 +52,35 @@ def editPost(id):
         db.session.commit()
         flash("Post ha sido actualizado")
         return redirect(url_for('public.post', id=id))
-    else:
+    if current_user.id == post.poster_id:
         form.title.data=post.title
+        # form.author.data=post.author
         form.slug.data=post.slug
         form.content.data=post.content
-        return render_template('public/editPost.html',form=form)
+        # return redirect(url_for('public.post', id=id, form=form))
+        return render_template('public/editPost.html', form=form)
+    else:
+        flash("You Aren't Authorized to Edit This Post...")
+        return redirect(url_for('public.post', id=id))
 @public_bp.route("/deletePost<int:id>")
+@login_required
 def deletePost(id):
     postDelete=Posts.query.get_or_404(id)
-    try:
-        db.session.delete(postDelete)
-        db.session.commit()
-        return redirect(url_for('public.index'))
-    except:
-            flash("Whoops! There was a problem deleting post")      
-            #Grab all the posts from the databases:
-            posts = Posts.query.order_by(Posts.date_posted)
-            return render_template("public/posts.html", posts= posts)
+    ide=current_user.id
+    if ide== postDelete.poster.id:       
+        try:
+            db.session.delete(postDelete)
+            db.session.commit()
+            flash("¡Post fue eliminado...!")
+            return redirect(url_for('public.index'))
+        
+        except:
+                flash("Whoops! There was a problem deleting post")      
+                #Grab all the posts from the databases:
+                posts = Posts.query.order_by(Posts.date_posted)
+                return render_template("public/posts.html", posts= posts)
+    else:
+        flash("You Aren't Authorized to Delete This Post")
+        return redirect(url_for('public.post', id=id))
 #https://www.youtube.com/shorts/1bX0xVcoX5w
